temporary backup of where I am currently at. what I'm trying to achieve: - have separate declaration files for inline props and symbolic props. - make both props mapped types compatible with each other. - make them both provide factory types for creating `InterinsicElements`, based on the user's provided framework types - `HyperRender` should now have a constructor that takes an optional config parameter, and all frameworks should accept the `config.hyperscript: s HyperZone` optional constructor parameter in order to either obtain a pre-existing framework-specific context or sideload their own into the `HyperZone` - remove the class factory construction of the tsignal reactive component renderer classes, and replace them with a class that accepts the hyperzone as an optional parameter to load its `tsignal.Context` from.
