+++ src/dom_core/mod.ts
@@ -1,7 +1,5 @@
 export type { HTMLElementUniqueMemberKeys, HTMLEventNames, HTMLTagNames, SVGEventNames, SVGTagNames } from "../deps.ts"
 export type { HTMLTagNameAttributesMap } from "./html_attributes_map.ts"
-export { inlinePropsRemapper } from "./inline_convenience.ts"
-export type { InlineAttrName, InlineEventName, InlineExecuteFn, InlineExecuteName, InlineMemberName } from "./inline_convenience.ts"
 export type { SVGTagNameAttributesMap } from "./svg_attributes_map.ts"
 import type { HTMLTagNames, SVGTagNames } from "../deps.ts"
 import type { IntrinsicElements } from "../tsignal/mod.ts"
